Accordingly to Issue #5

+++ bq769x0.cpp
@@ -620,7 +620,7 @@ void bq769x0::updateTemperatures()
 
 void bq769x0::updateCurrent(bool ignoreCCReadyFlag)
 {
-  int adcVal = 0;
+  int16_t adcVal = 0;
   regSYS_STAT_t sys_stat;
   sys_stat.regByte = readRegister(SYS_STAT);
